+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- **BREAKING**: Change `moveForward` to `moveRight`, `moveBackward` to `moveLeft`
+- **BREAKING**: Change `clearCurrentToEndScreen` to `clearCurrentToScreenEnd` (the same for other similar methods)
+
 ## 0.1.0
 
 - Initial version.